requirements.txt
asgiref==3.5.2; python_version >= "3.8"
django-filter==22.1; python_version >= "3.7"
django==4.1.4; python_version >= "3.8"
djangorestframework==3.14.0; python_version >= "3.6"
importlib-metadata==5.1.0; python_version < "3.10" and python_version >= "3.7"
markdown==3.4.1; python_version >= "3.7"
pytz==2022.6; python_version >= "3.6"
sqlparse==0.4.3; python_version >= "3.8"
tzdata==2022.7; sys_platform == "win32" and python_version >= "3.8"
zipp==3.11.0; python_version < "3.10" and python_version >= "3.7"
